fix: refactored all the sso backend code to use idiomatic functions and handle login and google logic properly (was majorly broken before)
